Split render by blocks
