+++ lib/src/render/blocks/heading.dart
@@ -1,0 +1,94 @@
+//ignore_for_file: unnecessary_import
+
+import 'dart:ui';
+
+import 'package:flutter/cupertino.dart';
+import 'package:flutter/gestures.dart';
+import 'package:flutter/material.dart';
+import 'package:flutter/rendering.dart';
+
+import '../../nodes.dart';
+import '../../theme.dart';
+import '../block_painter.dart';
+import '../span_builder.dart';
+
+/// A class for painting a heading block in markdown.
+class BlockPainter$Heading
+    with ParagraphGestureHandler, SelectableTextBlock
+    implements BlockPainter {
+  @override
+  TextPainter get selectionPainter => painter;
+
+  /// Creates a heading painter for [spans] at [level] (1-6), styled by [theme].
+  BlockPainter$Heading({
+    required int level,
+    required List<MD$Span> spans,
+    required this.theme,
+  }) : painter = TextPainter(
+          text: paragraphFromMarkdownSpans(
+            spans: spans,
+            theme: theme,
+            textStyle: theme.headingStyleFor(level),
+          ),
+          textAlign: TextAlign.start,
+          textDirection: theme.textDirection,
+          textScaler: theme.textScaler,
+        );
+
+  /// The theme used to style the heading.
+  final MarkdownThemeData theme;
+
+  /// The text painter that owns the heading's glyphs.
+  final TextPainter painter;
+
+  @override
+  Size get size => _size;
+  Size _size = Size.zero;
+
+  /// Last span hit by the tap down event.
+  TextSpan? _lastSpan;
+
+  @override
+  void handleTapDown(PointerDownEvent event) {
+    _lastSpan = null; // Reset the span on tap down.
+    final span = hitTestInlineSpanWithPointerEvent(event, painter);
+    if (span case TextSpan textSpan) _lastSpan = textSpan;
+  }
+
+  @override
+  void handleTapUp(PointerUpEvent event) {
+    if (_lastSpan == null) return; // No span was hit on tap down.
+    final span = hitTestInlineSpanWithPointerEvent(event, painter);
+    if (span != null && _lastSpan == span) {
+      // If the span is the same as the one hit on tap down,
+      // call the tap recognizer.
+      if (span case TextSpan(recognizer: TapGestureRecognizer(:var onTap)))
+        onTap?.call();
+    }
+    _lastSpan = null; // Clear the span after handling the tap.
+  }
+
+  @override
+  Size layout(double width) {
+    painter.layout(
+      minWidth: 0,
+      maxWidth: width,
+    );
+    return _size = painter.size;
+  }
+
+  @override
+  void paint(Canvas canvas, Size size, double offset) {
+    // If the width is less than required do not paint anything.
+    if (size.width < _size.width) return;
+    painter.paint(
+      canvas,
+      Offset(0, offset),
+    );
+  }
+
+  @override
+  void dispose() {
+    painter.dispose();
+  }
+}
